==> scripts/07_fun_K_mst.R <==
library(raster)
library(gdistance)
library(yenpathy)
library(rgdal)
library(sf)
library(magrittr)
library(matrixStats)


gen_top_tree <- function(tr, resist,  numpath, bufdist, pts){
  originCells <- raster::cellFromXY(tr, pts)
  goalCells <- raster::cellFromXY(tr, pts)
  indexOrigin <- originCells 
  indexGoal <- goalCells 
  result.list <- vector("list",numpath)
  update.res.list <- vector("list",numpath)
  for(z in 1:numpath){
    if(z == 1){
      y <- transitionMatrix(tr)
      if(isSymmetric(y)) {
        mode <- "undirected"
      }else{
        mode <- "directed"
      }
      adjacencyGraph <- igraph::graph.adjacency(y, mode=mode, weighted=TRUE)
      E(adjacencyGraph)$weight <- 1/E(adjacencyGraph)$weight
      dst <- distances(adjacencyGraph, indexOrigin, indexGoal, mode="all", weights =E(adjacencyGraph)$weight, algorithm = 'dijkstra')
      diag(dst) <- NA
      mst.grph <- graph.adjacency(dst, mode="undirected", weighted = TRUE, diag = FALSE)
      ms.tree <- mst(mst.grph, weights = E(mst.grph)$weight)
      pos <- get.edgelist(ms.tree)
      path.list <- vector("list",nrow(pos))
      for (i in 1:nrow(pos)){
        sp <- shortest_paths(adjacencyGraph, from = indexOrigin[pos[i,1]], to = indexGoal[pos[i,2]], mode = "out")
        path.list[i] <-sp$vpath 
      }
      #path.vect <- unlist(path.list)
      result <- tr
      transitionMatrix(result) <- Matrix::Matrix(0, ncol=ncell(tr), nrow=ncell(tr))
      for(i in 1:length(path.list)){
        sPVector <- path.list[[i]]
        adj <- cbind(sPVector[-(length(sPVector))], sPVector[-1])
        adj <- rbind(adj,cbind(adj[,2], adj[,1]))
        transitionMatrix(result)[adj] <- 1/length(path.list) + transitionMatrix(result)[adj]
      }
      result.rast <- raster(result)
      result.buf <- raster::buffer(result.rast, width = bufdist, doEdge=TRUE)
      result.list[[z]] <- result.buf
      result.buf.2 <- raster::buffer(result.rast, width = 2 * bufdist, doEdge=TRUE)
      update <- tr
      adj <- raster::adjacent(result.buf.2, Which(!is.na(result.buf.2), cells = TRUE), directions=16)
      transitionMatrix(update)[adj] <- cellStats(raster(tr), min)
      update.res.list[[z]] <- update
    }else{
      y <- transitionMatrix(update.res.list[[z-1]])
      if(isSymmetric(y)) {
        mode <- "undirected"
      }else{
        mode <- "directed"
      }
      adjacencyGraph <- igraph::graph.adjacency(y, mode=mode, weighted=TRUE)
      E(adjacencyGraph)$weight <- 1/E(adjacencyGraph)$weight
      dst <- distances(adjacencyGraph, indexOrigin, indexGoal, mode="all", weights =E(adjacencyGraph)$weight, algorithm = 'dijkstra')
      diag(dst) <- NA
      mst.grph <- graph.adjacency(dst, mode="undirected", weighted = TRUE, diag = FALSE)
      ms.tree <- mst(mst.grph, weights = E(mst.grph)$weight)
      pos <- get.edgelist(ms.tree)
      path.list <- vector("list",nrow(pos))
      for (i in 1:nrow(pos)){
        sp <- shortest_paths(adjacencyGraph, from = indexOrigin[pos[i,1]], to = indexGoal[pos[i,2]], mode = "out")
        path.list[i] <-sp$vpath 
      }
      result <- tr
      transitionMatrix(result) <- Matrix::Matrix(0, ncol=ncell(tr), nrow=ncell(tr))
      for(i in 1:length(path.list)){
        sPVector <- path.list[[i]]
        adj <- cbind(sPVector[-(length(sPVector))], sPVector[-1])
        adj <- rbind(adj,cbind(adj[,2], adj[,1]))
        transitionMatrix(result)[adj] <- 1/length(path.list) + transitionMatrix(result)[adj]
      }
      result.rast <- raster(result)
      result.buf <- raster::buffer(result.rast, width = bufdist, doEdge=TRUE)
      result.list[[z]] <- result.buf
      result.buf.2 <- raster::buffer(result.rast, width = 2 * bufdist, doEdge=TRUE)
      update <- update.res.list[[z-1]] 
      adj <- raster::adjacent(result.buf.2, Which(!is.na(result.buf.2), cells = TRUE), directions=16)
      transitionMatrix(update)[adj] <- cellStats(raster(tr), min)
      update.res.list[[z]] <- update
    }
  }
  all.res <- list(result.list, update.res.list)
  return(all.res)
}
